Add better support for Generics - also fix some issues with the filter envelope producing crap data

+++ filterapplier.go
@@ -7,5 +7,5 @@ import (
 // FilterApplier is an interface for applying a filter to a sample stream
 type FilterApplier interface {
 	ApplyFilter(dry volume.Matrix) volume.Matrix
-	SetFilterEnvelopeValue(envVal float32)
+	SetFilterEnvelopeValue(envVal int8)
 }
